"Use" the built-in engine version check
Engine knows better, no need to rely on string parsing hax.

It wasn't actually used, but I updated the commented out lines
since they serve as a reminder that this function exists.

The engine requires the middle value to be zero at the moment,
I am not sure why but it may well be a feature (since that value
is always 1 in builds and is pretty much redundant).